pyproject.toml
[project]
name = "via-archief"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "pillow>=11.0.0",
    "quart-db[sqlite]>=0.8.3",
    "quart>=0.19.8",
    "requests>=2.32.3",
]
